# [CI] Include maven ERROR messages in mandrel-integration-tests logs